requirements.txt
# panda3d 1.10.14
# python 3.10


panda3d-gltf
git+https://github.com/prasanthcodes/panda3d-simplepbr
gltflib


